+++ releasenotes.md
@@ -12,7 +12,7 @@
   - 居中不正确
   - 设置宽度无效
   - 设置高度无效
-  - 见 [流式段落布局示例](./ofdrw-layout/src/test/java/org/ofdrw/layout/ParagraphLayoutDemo.java)
+  - 见 [段落布局示例](./ofdrw-layout/src/test/java/org/ofdrw/layout/cases/content/ParagraphCase.java)
 
 ## Version 1.9.5 2021-5-24 20:52:26